Jochong (#11)
0.0.1

+++ module/channelControl.js
@@ -1,10 +1,12 @@
 const elastic = require('elasticsearch');
-const { Client, Pool } = require('pg');
+const { Client } = require('pg');
 const redis = require('redis').createClient();
 const pgConfig = require('../config/psqlConfig');
 const channelDataValidCheck = require('./channelDataValidCheck');
 const passwordHash = require('./passwordHash');
 const verifyToken = require('../module/verifyToken');
+const AWS = require('aws-sdk');
+const awsConfig = require('../config/awsConfig');
 
 const addChannel = (channelData, loginType = 'local') => {
     return new Promise(async (resolve, reject) => {
@@ -18,6 +20,9 @@ const addChannel = (channelData, loginType = 'local') => {
         }
 
         const pgClient = new Client(pgConfig);
+        const esClient = elastic.Client({
+            node : "http://localhost:9200"
+        });
 
         try{
             //connect psql
@@ -108,17 +113,13 @@ const addChannel = (channelData, loginType = 'local') => {
                 //INSERT psql
                 await pgClient.query(insertSql, insertArray);
 
-                //index elasticsearch
-                const esClient = elastic.Client({
-                    node : "http://localhost:9200"
-                });
                 await esClient.index({
                     index : 'channel',
                     id : email,
                     body : {
                         name : channelName
                     }
-                })
+                });
             }
             
             //COMMIT
@@ -142,13 +143,13 @@ const addChannel = (channelData, loginType = 'local') => {
     })
 }
 
-const getChannel = (channelEamil) => {
+const getChannel = (channelEamil, loginUserEmail = '') => {
     return new Promise(async (resolve, reject) => {
         const pgClient = new Client(pgConfig);
         try{
             await pgClient.connect();
 
-            const selectChannelSql = 'SELECT name, birth, sex, authority, creation_time, description, profile_img FROM shoot.channel WHERE email = $1';
+            const selectChannelSql = 'SELECT email, name, birth, sex, authority, creation_time, description, profile_img, subscribe_count FROM shoot.channel WHERE email = $1';
             const selectChannelResult = await pgClient.query(selectChannelSql, [channelEamil]);
 
             if(selectChannelResult.rows.length === 0){
@@ -159,16 +160,15 @@ const getChannel = (channelEamil) => {
             }else{
                 const channelData = selectChannelResult.rows[0];
 
-                resolve({
-                    email : channelEamil,
-                    name : channelData.name,
-                    birth : channelData.birth,
-                    sex : channelData.sex,
-                    authority : channelData.int,
-                    creation_time : channelData.creation_time,
-                    description : channelData.description,
-                    profile_img : channelData.profile_img
-                });
+                if(loginUserEmail.length !== 0){
+                    //SELECT subscribe
+                    const selectSubscribeSql = 'SELECT * FROM shoot.subscribe WHERE subscriber_channel_email = $1 AND subscribed_channel_email = $2';
+                    const selectSubscribeResult = await pgClient.query(selectSubscribeSql, [loginUserEmail, channelEamil]);
+
+                    channelData.subscribe_state = selectSubscribeResult.rows[0] !== undefined;
+                }
+
+                resolve(channelData);
             }
         }catch(err){
             reject({
@@ -182,7 +182,6 @@ const getChannel = (channelEamil) => {
 
 const getAllChannel = (searchKeyword, scrollId = undefined, size = 30) => {
     return new Promise(async (resolve, reject) => {
-        //connect es
         const esClient = new elastic.Client({
             node : 'http://localhost:9200'
         })
@@ -254,25 +253,177 @@ const getAllChannel = (searchKeyword, scrollId = undefined, size = 30) => {
     })
 }
 
+const modifyChannel = (loginUserEmail, channelData) => {
+    return new Promise(async (resolve, reject) => {
+        AWS.config.update(awsConfig);
+        const s3 = new AWS.S3();
+        const pgClient = new Client(pgConfig);
+        const esClient = new elastic.Client({
+            node : 'http://localhost:9200'
+        });
+
+        const modifyChannelImg = channelData.channelImg || '';
+        const modifyChannelName = channelData.channelName || '';
+        const modifyDescription = channelData.description || '';
+        const modifyBirth = channelData.birth;
+        const modifySex = channelData.sex;
+
+        try{
+            await pgClient.connect();
+
+            //BEGIN
+            await pgClient.query('BEGIN');
+
+            //SELECT profile img 
+            const selectChannelSql = 'SELECT profile_img FROM shoot.channel WHERE email = $1';
+            const selectChannelResult = await pgClient.query(selectChannelSql, [loginUserEmail]);
+
+            //UPDATE name,sex,birth,description
+            const updateObj = {};
+            if(modifyChannelImg.length === 0){
+                updateObj.sql = `UPDATE
+                                            shoot.channel
+                                        SET
+                                            name = $1,
+                                            sex = $2,
+                                            birth = $3,
+                                            description = $4
+                                        WHERE
+                                            email = $5
+                                        `;
+                updateObj.dataArray = [modifyChannelName, modifySex, modifyBirth, modifyDescription, loginUserEmail];
+            }else{
+                updateObj.sql = `UPDATE
+                                    shoot.channel
+                                SET
+                                    name = $1,
+                                    sex = $2,
+                                    birth = $3,
+                                    description = $4,
+                                    profile_img = $5
+                                WHERE
+                                    email = $6
+                                `;
+                updateObj.dataArray = [modifyChannelName, modifySex, modifyBirth, modifyDescription, modifyChannelImg, loginUserEmail];
+            }
+            await pgClient.query(updateObj.sql, updateObj.dataArray);
+            
+            //delete channel profile img on s3
+            if(selectChannelResult.rows[0].profile_img && modifyChannelImg){
+                await s3.deleteObject({
+                    Bucket: 'jochong/channel_img', 
+                    Key: selectChannelResult.rows[0].profile_img
+                }).promise();
+            }
+
+            //update es
+            await esClient.index({
+                index : 'channel',
+                id : loginUserEmail,
+                body : {
+                    name : modifyChannelName
+                }
+            });
+
+            //COMMIT
+            await pgClient.query('COMMIT');
+
+            resolve(1);
+        }catch(err){
+            //ROLLBACK
+            await pgClient.query('ROLLBACK');
+
+            reject({
+                statusCode : 409,
+                message : 'unexpected error occured',
+                err : err
+            });
+        }
+    });
+}
+
+const modifyPw = (loginUserEmail = '', pwData = {}) => {
+    return new Promise(async (resolve, reject) => {
+        const pgClient = new Client(pgConfig);
+
+        const modifyPw = pwData.pw;
+        const modifyPwCheck = pwData.pwCheck
+
+        //pw valid check
+        const pwExp = new RegExp('^(?=.*[a-zA-Z])(?=.*[0-9])(?=.*[!@#$%^&*])[a-zA-Z0-9!@#$%^&*]{8,16}$');
+        if(!pwExp.test(modifyPw)){
+            reject({
+                statusCode : 400,
+                message : "invalid password"
+            });
+            return;
+        }
+
+        //pwCheck valid check
+        if(modifyPw !== modifyPwCheck){
+            reject({
+                statusCode : 400,
+                message : "invalid password check"
+            });
+            return;
+        }
+
+        try{
+            await pgClient.connect();
+
+            await redis.connect();
+
+            const authState = await redis.exists(`certified_email_${loginUserEmail}`);
+
+            if(authState){
+                //UPDATE channel
+                const updateSql = 'UPDATE shoot.channel SET pw = $1 WHERE email = $2';
+                await pgClient.query(updateSql, [passwordHash(modifyPw), loginUserEmail]);
+
+                await redis.del(`certified_email_${loginUserEmail}`);
+
+                await redis.disconnect();
+
+                resolve(1);
+            }else{
+                await redis.disconnect();
+                
+                reject({
+                    statusCode : 403,
+                    message : 'no auth'
+                });
+            }
+        }catch(err){
+            reject({
+                statusCode : 409,
+                message : 'unexpected error occureed',
+                err : err
+            });
+        }
+    });
+}
+
 const deleteChannel = (deleteEmail, token) => {
     return new Promise(async (resolve, reject) => {
+        AWS.config.update(awsConfig);
+        const s3 = new AWS.S3();
+        const pgClient = new Client(pgConfig);
+        const esClient = elastic.Client({
+            node : "http://localhost:9200"
+        });
+
         const verify = verifyToken(token);
 
         if(verify.state){
             const loginUserEmail = verify.email;
+            const loginUserAuthority = verify.authority;
             try{
-                const pgClient = new Client(pgConfig);
                 await pgClient.connect();
 
-                const esClient = elastic.Client({
-                    node : "http://localhost:9200"
-                });
-
                 //SELECT login user email and authority
-                const selectSql = 'SELECT email, authority FROM shoot.channel WHERE email = $1';
+                const selectSql = 'SELECT email, authority, profile_img FROM shoot.channel WHERE email = $1';
                 const selectResult = await pgClient.query(selectSql, [deleteEmail]);
 
-                console.log(deleteEmail);
                 //check email existence
                 if(selectResult.rows.length === 0){
                     reject({
@@ -282,15 +433,27 @@ const deleteChannel = (deleteEmail, token) => {
                     return;
                 }
 
+                //check delete channel authority
+                if(selectResult.rows[0].authority === 1){
+                    reject({
+                        message : 'cannot delete admin account',
+                        auth : 403
+                    });
+                    return;
+                }
+
                 //check authority
-                const loginUserAuthority = selectResult.rows[0].authority;
                 if(loginUserEmail === deleteEmail || loginUserAuthority === 1){
                     //BEGIN
                     await pgClient.query('BEGIN');
 
-                    //DELETE post
-                    const updatePostQuery = 'UPDATE shoot.post SET delete_time = $1 WHERE upload_channel_email = $2';
-                    await pgClient.query(updatePostQuery, [deleteEmail]);
+                    if(selectResult.rows[0].profile_img){
+                        //delete channel profile img on s3
+                        await s3.deleteObject({
+                            Bucket: 'jochong/channel_img', 
+                            Key: selectResult.rows[0].profile_img
+                        }).promise();
+                    }
 
                     //delete post on elasticsearch
                     await esClient.deleteByQuery({
@@ -304,20 +467,27 @@ const deleteChannel = (deleteEmail, token) => {
                         }
                     });
 
-                    //DELETE comment
-                    const deleteCommentQuery = 'DELETE FROM shoot.comment WHERE wirte_channel_email = $1';
-                    const deleteCommentResult = await pgClient.query(deleteCommentQuery, [deleteEmail]);
-                    console.log(deleteCommentResult.rows);
-
-                    //DELETE channel
-                    const deleteChannelQuery = 'DELETE FROM shoot.channel WHERE email = $1';
-                    await pgClient.query(deleteChannelQuery, [deleteEmail]);
+                    //delete comment on elasticsearch
+                    await esClient.deleteByQuery({
+                        index : 'comment',
+                        body : {
+                            query : {
+                                match : {
+                                    write_channel_email : deleteEmail
+                                }
+                            }
+                        }
+                    });
                     
                     //delete channel on elasticsearch
                     await esClient.delete({
                         index : 'channel',
                         id : deleteEmail
                     });
+
+                    //DELETE channel
+                    const deleteChannelQuery = 'DELETE FROM shoot.channel WHERE email = $1';
+                    await pgClient.query(deleteChannelQuery, [deleteEmail]);
 
                     //COMMIT
                     await pgClient.query('COMMIT');
@@ -330,6 +500,8 @@ const deleteChannel = (deleteEmail, token) => {
                     })
                 }
             }catch(err){
+                await pgClient.query('ROLLBACK');
+
                 reject({
                     message : 'unexpected error occured',
                     statusCode : 409,
@@ -349,5 +521,7 @@ module.exports = {
     addChannel : addChannel,
     getChannel : getChannel,
     getAllChannel : getAllChannel,
-    deleteChannel : deleteChannel
+    deleteChannel : deleteChannel,
+    modifyChannel : modifyChannel,
+    modifyPw : modifyPw
 }
